services/control-service/requirements.txt
fastapi
uvicorn[standard]
kafka-python
paho-mqtt
prometheus-client
python-dotenv